refactor(material/core): remove underscores from symbols consumed in other files (#20906)
In the new Sass module system a symbol prefixed with an underscore means that it's
private and it can't be used in other files. These changes refactor all of our existing
usages to remove the underscore. Some of the symbols have been moved into
`private` directories to make it more obvious that they shouldn't be depended upon.