fix(cdk/drag-drop): don't block scrolling if it happens before delay has elapsed (#21382)
The `DragRef` has the ability to add a delay to the dragging so that scrolling isn't
blocked on touch devices. Currently this doesn't really work, because we always
block the native event as a result of a949db30c7554c32df77b81d163fe7d8cbb83c97.
These changes move some code around so that we only block the events if the
item considers itself as being in a "dragged" state (the delay has elapsed and the
user is past the drag threshold).

Fixes #17923.