feat(cdk-experimental/table-scroll-container): Create directive and demo (#21200)
Adds the CdkTableScrollContainer and some hooks in CdkTable for it.
For Webkit/Blink browsers, this sizes the scroll bars in coordinating with
sticky rows/columns in the table to create the user experience that a lot of
commenters in #5885 seemed to want but without the accessibility problems of
other approaches.