fix(material/table): increase specificity on sticky selector (#21403)
The sticky styles were switched to a class in #19823 which has a very low specificity.
As a result, there's a high chance of it being overwritten by some user styles.

These changes use `!important` so that we get a similar specificity as when the styles
were being set inline before #19823.

Fixes #20558.