fix(cdk/overlay): sub-pixel deviations throwing off positioning in some cases (#21427)
If the browser is zoomed in beyond the default level, it may report numbers with
sub-pixel deviations to `getBoundingClientRect` (e.g. 100.09 vs 100) which can throw off
our logic when comparing against the viewport size which is always a whole number.

These changes fix the issue by rounding down the numbers that we get from `getBoundingClientRect`.

Fixes #21350.