fix(material/schematics): don't drop imports in files that do not use theming APIs (#22438)
I noticed this while trying out the theming migration against the docs site. The way the `theming-api` migration is set up means that we'll drop the Material imports as soon as we see them, but that doesn't guarantee that they're actually used further down in the file.

These changes rework the logic so that we drop the imports after we've migrated the symbols.