fix(cdk/testing): unable to assign value to number inputs using sendKeys (#22395)
Currently the `UnitTestElement` simulates typing into an input by assigning the value character-by-character and dispatching fake events along the way. The problem is that for input types that require the value to be in a particular format (e.g. `number`, `color`, `date`) doing so will temporarily assign an invalid value which will be rejected by the browser with a warning like `The specified value "12." cannot be parsed, or is out of range.`.

This can become a problem for some common use cases like the `ReactiveFormsModule` where a directive might be keeping track of the value by looking at the DOM inside of an `input` event (e.g. the `FormControl` directive does this).

These changes resolve the issue by looking at the type of the input, and if it's a type that requires a specific format, we assign the value immediately.

Fixes #22129.