feat(material/tabs): allow for content tabindex to be customized (#21912)
In some cases, adding a `tabindex` to a `tabpanel` can improve accessibility, however that
is currently impossible with our setup, because the element isn't exposed. These changes
add an input that allows consumers to customize the `tabindex` based on their use case.

We tried to tackle this in an earlier PR (#14808), but it was decided not to proceed,
because the `tabindex` isn't relevant for all use cases. This approach should be a bit
more flexible since it allows users to opt into it.

Fixes #21819.