fix(material/schematics): theming migration schematic not inserting use statement at the top of the file in some cases
Our logic for determining where to insert an `@use` statement is to look for the first `@import` and insert before it. The problem is that if the theming import is the only one in the file and it's further down, we'll insert it incorrectly, because `@use` has to always be above other statements.