fix(material-experimental/mdc-list): error with latest MDC list canary version (#22636)
MDC updated the focus strategy for lists with multi-selection enabled.
This breaks our tests as we relied on the old strategy where the
previously focused item always becomes the focus target.

This is no longer the case as MDC now intends to focus the first
selected item of a list instead (except for non-selection lists which
are interactive; here the previously focused item is still used).

We update the tests and update the selection list so that pre-selected
options are recognized correctly so that MDC's new focus strategy works
as expected. Before this change we didn't sync the pre-selected options
with the MDC list foundation.