fix(material/datepicker): pass correct inject flags to injector (#22665)
We have a call to `Injector.get` in the date range input which is supposed to fall back to `null` if the value doesn't resolve, however we're missing the `Optional` flag. It works by accident at the moment, because the inject flags aren't actually being passed through by the framework.

https://github.com/angular/angular/pull/41592 is supposed to fix the framework issue, but their CI will break due to our usage.

These changes fix our usage in order to unblock the framework PR.