docs(material/autocomplete): add descriptions for properties showing up in API docs (#22851)
Currently the `options` and `optionGroups` properties are showing up
in the docs even though they are marked as `@docs-private` in the
base class. This happens because they are declared in the derived
autocomplete implementations (both non-MDC and MDC). Given these
properties not being prefixed with an underscore anyway, we should
start documenting these.