build: ampersand selector lint rule not catching some cases (#22974)
I noticed this while working on something else. The ampersand selector mixin looks for one `&` instance and then gives up, but there might be other ampersands inside the same selector which are invalid.

These changes update the logic to be a bit more robust and remove the limitation where failures weren't reported on private mixins.