fix(material/tabs): update tab nav bar focused index on direct focus (#22584)
The tab nav bar allows navigation both through tabbing and keyboard events. This introduces a problem when the user tabs into it and then starts using the arrow keys, because the `ListKeyManager` state isn't up-to-date.

These changes add some logic to update the index when the tabs receives focus.

Fixes #22576.